requirements.txt
Flask==2.2.2
PyPDF2==3.0.1
nltk==3.8.1
gensim==3.6.0
spacy==3.5.1
pymongo==4.4.0
concurrent-log-handler==0.9.22
mongomock==4.1.2
Werkzeug==2.2.2
rake-nltk==1.0.6
python-dotenv==1.0.1
pytesseract==0.3.10
# en_core_web_sm==3.5.0
scikit-learn==1.5.2
PyMuPDF==1.24.11
python-docx==1.1.2